+++ api/src/Communication/Service/InternalMessageManager.php
@@ -138,15 +138,39 @@ class InternalMessageManager
     }
 
     /**
-     * Get a complete message
+     * Get a complete message thread
+     * @param int $idMessage    The message we want the thread
+     * @param bool $checkRead   If true, we check the current message as read
+     * @param int $userId       Id of the requester. Usefull if checkRead is true
      */
-    public function getCompleteThread($idMessage)
+    public function getCompleteThread(int $idMessage, bool $checkRead=false, int $userId=null)
     {
         $message = $this->messageRepository->find($idMessage);
         if (empty($message)) {
             throw new MessageNotFoundException("Message not found");
         }
-        return  array_merge([$message], $message->getMessages());
+        $messages = array_merge([$message], $message->getMessages());
+        
+        // getCompleteThread is called in various ways that does'nt require that the read status be updated.
+        // For example in, UserManager -> getProfileSummary
+        if ($checkRead) {
+            foreach ($messages as $currentMessage) {
+                foreach ($currentMessage->getRecipients() as $recipient) {
+                    if (is_null($userId)) {
+                        throw new \LogicException("No user specified");
+                    }
+                    $userRecipientId = $recipient->getUser()->getId();
+                    // We set a read date only if the recipient userid is the requester id
+                    if ($userId == $userRecipientId) {
+                        $recipient->setReadDate(new \DateTime("now"));
+                    }
+                    $this->entityManager->persist($currentMessage);
+                }
+            }
+            $this->entityManager->flush();
+        }
+
+        return array_merge([$message], $message->getMessages());
     }
 
     public function getMessage($idMessage)